requirements.txt
asgiref==3.5.2
distlib==0.3.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
filelock==3.7.1
Markdown==3.4.1
mysqlclient==2.1.1
platformdirs==2.5.2
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.3
django-cors-headers ==3.13.0
requests == 2.28.1